fix : github action 에러 수정
